LoanController tests
+added tests for LoanController::getAllLoans
+added tests for LoanController::createLoan

+++ SmoothBankCard/src/main/java/com/ss/sbank/dao/loans/LoanDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ss.sbank.de.loans.Loan;
 
-public interface LoanDAO extends JpaRepository<Loan, Integer> {
+public interface LoanDAO extends JpaRepository<Loan, Long> {
 
 }
